added configurable params for deepab example
