+++ examples/cnn/deeplab/train.py
@@ -1,3 +1,6 @@
+import os
+import argparse
+
 import numpy as np
 from sklearn.utils import shuffle
 from sklearn.metrics import confusion_matrix
@@ -5,7 +8,14 @@ from neupy import algorithms, storage
 
 from data import get_training_data, get_validation_data
 from model import create_deeplab_model
+from utils import score_segmentation
 from resnet50 import download_resnet50_weights
+
+
+parser = argparse.ArgumentParser()
+parser.add_argument('--storage-folder', '-s', required=True)
+parser.add_argument('--epochs', '-e', type=int, default=30)
+parser.add_argument('--batch-size', '-b', type=int, default=10)
 
 
 def iter_batches(images, annotations, batch_size=10):
@@ -16,38 +26,12 @@ def iter_batches(images, annotations, batch_size=10):
         yield images[i:i + batch_size], annotations[i:i + batch_size]
 
 
-def get_confusion_matrix(y_val, y_pred):
-    y_valid_label = y_val.max(axis=-1)
-    interested_labels = y_valid_label == 1
-
-    y_expected = y_val.argmax(axis=-1)[interested_labels]
-    y_pred = y_pred.argmax(axis=-1)[interested_labels]
-
-    labels = list(range(21))
-    return confusion_matrix(y_expected, y_pred, labels=labels)
-
-
-def segmentation_metrics(confusion):
-    tp = np.diag(confusion)
-    sum_cols = confusion.sum(axis=1)
-    sum_rows = confusion.sum(axis=0)
-
-    union = (sum_cols + sum_rows - tp)
-    iou = tp / union
-    miou = np.mean(iou[(union > 0) & (sum_cols > 0)])
-    accuracy = tp.sum() / confusion.sum()
-
-    return accuracy, miou
-
-
-def score_segmentation(y_val, y_pred):
-    confusion = get_confusion_matrix(y_val, y_pred)
-    return segmentation_metrics(confusion)
-
-
 if __name__ == '__main__':
-    batch_size = 10
-    n_epochs = 30
+    args = parser.parse_args()
+    storage_folder = args.storage_folder
+
+    if not os.path.exists(storage_folder):
+        os.mkdir(storage_folder)
 
     resnet50_weights_filename = download_resnet50_weights()
     resnet50, deeplab = create_deeplab_model(resnet50_weights_filename)
@@ -66,10 +50,10 @@ if __name__ == '__main__':
         decay_rate=0.0001,
     )
 
-    for i in range(n_epochs):
+    for i in range(args.epochs):
         print("Epoch #{}".format(i + 1))
 
-        for x_batch, y_batch in iter_batches(images, annotations, batch_size):
+        for x_batch, y_batch in iter_batches(images, annotations, args.batch_size):
             x_batch = resnet50.predict(x_batch)
             optimizer.train(x_batch, y_batch, epochs=1, summary='inline')
 
@@ -80,6 +64,8 @@ if __name__ == '__main__':
         print("Val accuracy: {:.3f}".format(accuracy))
         print("Val miou: {:.3f}".format(miou))
 
-        filename = 'dump_v7/deeplab_v07_{:0>3}_{:.3f}_{:.3f}.hdf5'.format(i, accuracy, miou)
-        print("Saved: {}".format(filename))
-        storage.save(deeplab, filename)
+        filename = 'deeplab_{:0>3}_{:.3f}_{:.3f}.hdf5'.format(i, accuracy, miou)
+        filepath = os.path.join(storage_folder, filename)
+
+        print("Saved: {}".format(filepath))
+        storage.save(deeplab, filepath)
